Update to rubocop 0.75, massive lib/ lint fixes.
This version of rubocop is more aggressive than the previous versions the
project was using. As it is based on the community driven Ruby Style guide, I've
tried to increase our adherence to the standard.

The most significant change is the addition of `frozen_string_literal: true` on
top of all the files, which required to update some of the code to get the tests
running. The changes weren't too important but it might impact things that
aren't unit tested. Fingers crossed

Signed-off-by: Julien 'Lta' BALLET <contact@lta.io>